+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.26.0 - TBD
+
+This release adds support for transcoding DBN data into Apache parquet.
+
+#### Enhancements
+- Added `DBNStore.to_parquet` for transcoding DBN data into Apache parquet using `pyarrow`
+
 ## 0.25.0 - 2024-01-09
 
 #### Breaking changes